+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1a3 (21.04.2020)
 ### Feature
 * added support for pages with section
+* added support for `ALTERNATEALIGNMENT` value
 
 ## 0.1a2 (21.04.2020)
 ### Feature